+++ RevitAddin.CommandBindingTester/Revit/Commands/CreateCommand.cs
@@ -2,14 +2,13 @@ using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using RevitAddin.CommandBindingTester.ViewModels;
-using RevitAddin.CommandBindingTester.Views;
 using System;
 using System.Linq;
 
 namespace RevitAddin.CommandBindingTester.Revit.Commands
 {
     [Transaction(TransactionMode.Manual)]
-    public class Command : IExternalCommand
+    public class CreateCommand : IExternalCommand
     {
         public Result Execute(ExternalCommandData commandData, ref string message, ElementSet elementSet)
         {
@@ -19,13 +18,12 @@ namespace RevitAddin.CommandBindingTester.Revit.Commands
             var commands = Enum.GetValues(typeof(PostableCommand))
                 .OfType<PostableCommand>();
 
-            var models = commands
-                .Select(x => Models.BindingModel.Create(x))
-                .OrderBy(x => x.Id)
-                .ToList();
+            //var models = commands
+            //    .Select(x => Models.CreateBindingModel.Create(x))
+            //    .OrderBy(x => x.Id)
+            //    .ToList();
 
-            var viewModel = new BindingViewModel();
-            viewModel.Models = new System.Collections.ObjectModel.ObservableCollection<Models.BindingModel>(models);
+            var viewModel = new CreateBindingViewModel();
             var result = viewModel.ShowDialog();
 
             return Result.Succeeded;