+++ RevitAddin.CommandBindingTester/Views/CreateBindingView.xaml.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Windows;
+
+namespace RevitAddin.CommandBindingTester.Views
+{
+    public partial class CreateBindingView : Window
+    {
+        public CreateBindingView()
+        {
+            InitializeComponent();
+            InitializeWindow();
+        }
+
+        #region InitializeWindow
+        private void InitializeWindow()
+        {
+            this.SizeToContent = SizeToContent.WidthAndHeight;
+            this.ShowInTaskbar = false;
+            this.ResizeMode = ResizeMode.NoResize;
+            this.WindowStartupLocation = WindowStartupLocation.CenterScreen;
+        }
+        #endregion
+    }
+}